yet another intermediate revision
